libinstaller: Explicit failure if path isn't writable
As per bug #4, we should report a clear failure if the target path is
not writable.

The current code was catching the fact the file was not writable
but it didn't wrote an explicit message and even more confusing, was
trying to process the file descriptor leading to a creepy "Bad file
descriptor" error message.

This patch does return EACCES to avoid process the file description.
It also print an explicit message saying a particular file isn't
writable which generates a fatal error.

A typical output looks like :
	[root@host]: extlinux --once=plop /boot
	/boot is device /dev/sda
	Cannot open file '/boot/ldlinux.sys' in read/write mode !
	Fatal error, exiting.