com32: Adding 'exit' entry in chain.c32
This patch is coming after some bad experience with gfxboot.
GFXboot was padding unexpected options on the chain.c32 commande line.
As a result, chain was confused and didn't provided the expected
behavior.

This patch is a workaround to enforce chain.c32 not considering anything
on the command line after this keyword.

Example:

chain.c32 hd0 exit ...garbage...