chain: Adding diskbypname= option
The exisiting label= option offer to boot on a gpt partition that have a particular
label.
The existing guuid= option offer to boot on a disk or partition with a
particular label.

This new option offer to boot the disk that have a partition which have a given
label. The label is so just a way to detect a disk to boot.

A typical use is for making an intelligent localboot like :

    label localboot
        com32 chain.c32
        append diskbypname=datapartition

This allow booting on a disk that sports a least one partition named "datapartition".

You can consider diskbypname= doing almost what mbr= does but by inspecting the gpt partitions name.